Ajustes na interface e controller

+++ spring04/src/main/java/br/itau/spring04/controller/ContaController.java
@@ -30,7 +30,7 @@ public class ContaController {
 
 
     @GetMapping("/contas/{numero}")
-	public ResponseEntity<Conta> recuperarPeloNumero(@PathVariable int numero) {
+	public ResponseEntity<Conta> recuperarPeloNumero(@PathVariable Long numero) {
 		Conta resultado = repo.findById(numero).orElse(null);
 
 		if (resultado != null) {
