+++ spring04/src/main/java/br/itau/spring04/repository/ContaRepo.java
@@ -6,6 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.itau.spring04.model.Conta;
 
-public interface ContaRepo extends CrudRepository<Conta, Integer>{
+
+public interface ContaRepo extends CrudRepository<Conta, Long>{
     public List<Conta> findAll();
 }
